Number inputs to have decimal steps, max, and min

+++ HTMLRenderingTool/src/forms/Number.java
@@ -12,12 +12,12 @@ public class Number extends Input
 	/**
 	 * Max attribute.
 	 */
-	private int max;
+	private double max;
 	
 	/**
 	 * Min attribute.
 	 */
-	private int min;
+	private double min;
 	
 	/**
 	 * Placeholder attribute.
@@ -27,7 +27,7 @@ public class Number extends Input
 	/**
 	 * Step attribute.
 	 */
-	private int step;
+	private double step;
 	
 	/**
 	 * Constructor. Creates an input of type "number".
@@ -52,7 +52,7 @@ public class Number extends Input
 	 * Gets the step attribute.
 	 * @return step value
 	 */
-	public int getStep()
+	public double getStep()
 	{
 		return step;
 	}
@@ -61,7 +61,7 @@ public class Number extends Input
 	 * Sets the step attribute. < 0 to remove.
 	 * @param step step value
 	 */
-	public void setStep(int step)
+	public void setStep(double step)
 	{
 		this.step = step < 0 ? -1 : step;
 		if(this.step < 0)
@@ -104,7 +104,7 @@ public class Number extends Input
 	 * Gets the min attribute.
 	 * @return min value
 	 */
-	public int getMin()
+	public double getMin()
 	{
 		return min;
 	}
@@ -113,7 +113,7 @@ public class Number extends Input
 	 * Sets the min attribute. < 0 to remove.
 	 * @param min min value
 	 */
-	public void setMin(int min)
+	public void setMin(double min)
 	{
 		this.min = min < 0 ? -1 : min;
 		if(this.min < 0)
@@ -130,7 +130,7 @@ public class Number extends Input
 	 * Gets the max attribute.
 	 * @return max value
 	 */
-	public int getMax()
+	public double getMax()
 	{
 		return max;
 	}
@@ -139,7 +139,7 @@ public class Number extends Input
 	 * Sets the max attribute. < 0 to remove.
 	 * @param max max value
 	 */
-	public void setMax(int max)
+	public void setMax(double max)
 	{
 		this.max = max < 0 ? -1 : max;
 		if(this.max < 0)
@@ -159,7 +159,7 @@ public class Number extends Input
 	 */
 	public void parseStep(String step)
 	{
-		setStep(step == null ? -1 : Integer.parseInt(step));
+		setStep(step == null ? -1 : Double.parseDouble(step));
 	}
 	
 	/**
@@ -169,7 +169,7 @@ public class Number extends Input
 	 */
 	public void parseMin(String min)
 	{
-		setMin(min == null ? -1 : Integer.parseInt(min));
+		setMin(min == null ? -1 : Double.parseDouble(min));
 	}
 	
 	/**
@@ -179,6 +179,6 @@ public class Number extends Input
 	 */
 	public void parseMax(String max)
 	{
-		setMax(max == null ? -1 : Integer.parseInt(max));
+		setMax(max == null ? -1 : Double.parseDouble(max));
 	}
 }
